requirements.txt
streamlit==1.38.0
openai==1.51.0
numpy==1.26.4
scipy==1.11.4
nibabel==5.2.1
plotly==5.22.0
pillow==10.4.0
trimesh==4.0.8
